Refactor: add validation and change ui
