+++ front/src/components/userRegister/resstPass/RessetPassword.scss
@@ -1,0 +1,68 @@
+@import "@/assets/fonts/fonts.css";
+
+.ressetpass-section {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+.ressetpass-title {
+  width: 100%;
+  height: 10%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  margin-bottom: 2%;
+
+  h1 {
+    font-family: bold;
+    font-size: 3em;
+    color: #db1349;
+    margin-bottom: 1%;
+  }
+
+  p {
+    text-decoration: none;
+    color: #fff;
+    font-family: medium;
+
+    font-size: 18px;
+  }
+}
+.ressetpass-form {
+  width: 100%;
+  height: 30%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+
+  input {
+    width: 220px;
+    height: 50px;
+    border-radius: 40px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    text-align: center;
+    font-family: medium;
+    margin-bottom: 2%;
+  }
+  input ::placeholder {
+    color: rgba(255, 255, 255, 0.917);
+    text-align: center;
+  }
+}
+
+.ressetpass-error {
+  width: 100%;
+  height: 10%;
+  color: #db1349;
+  font-family: medium;
+  font-size: 23px;
+}
